+++ tests/TestCase.php
@@ -3,6 +3,7 @@
 namespace Sfneal\Dependencies\Tests;
 
 use Illuminate\Foundation\Application;
+use Illuminate\Http\Client\Response;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Http;
 use Sfneal\Dependencies\Providers\DependenciesServiceProvider;
@@ -13,6 +14,11 @@ use Sfneal\Helpers\Strings\StringHelpers;
 
 abstract class TestCase extends \Orchestra\Testbench\TestCase
 {
+    /**
+     * @var int Number of packages to be return from the packageProvider
+     */
+    protected $numberOfPackages = 3;
+
     /**
      * Register package service providers.
      *
@@ -33,11 +39,26 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
      */
     public function packageProvider(): array
     {
-        return [
+        $packages = [
+            ['sfneal/actions'],
+            ['sfneal/controllers'],
             ['sfneal/laravel-helpers'],
+            ['sfneal/queueables'],
+            ['sfneal/redis-helpers'],
+            ['sfneal/scopes'],
+            ['sfneal/string-helpers'],
+            ['sfneal/time-helpers'],
+            ['sfneal/tracking'],
             ['symfony/console'],
             ['spatie/laravel-view-models'],
+            ['webmozart/assert'],
+            ['spatie/laravel-settings'],
+            ['illuminate/database'],
+            ['laravel/framework'],
         ];
+        shuffle($packages);
+
+        return $packages;
     }
 
     /**
@@ -73,116 +94,153 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
         $this->assertSame($expected, $collection->count());
 
         $collection->each(function (DependenciesService $service) {
-            $this->assertTravisSvg($service->packageGitubName, $service->travis());
-            $this->assertVersionSvg($service->package, $service->version());
-            $this->assertLastCommitSvg($service->packageGitubName, $service->lastCommit());
-            $this->assertGithubUrl($service->packageGitubName, $service->gitHub());
-            $this->assertTravisUrl($service->packageGitubName, $service->travis());
-            $this->assertVersionUrl($service->package, $service->version());
+            $this->assertTravisSvg($service->packageGitubName, $service->travis(), false);
+            $this->assertVersionSvg($service->package, $service->version(), false);
+            $this->assertLastCommitSvg($service->packageGitubName, $service->lastCommit(), false);
+            $this->assertGithubUrl($service->packageGitubName, $service->gitHub(), false);
+            $this->assertTravisUrl($service->packageGitubName, $service->travis(), false);
+            $this->assertVersionUrl($service->package, $service->version(), false);
         });
     }
 
     /**
      * @param string $package
      * @param DependencySvg $generator
+     * @param bool $sendRequest
      */
-    public function assertTravisSvg(string $package, DependencySvg $generator)
+    public function assertTravisSvg(string $package, DependencySvg $generator, bool $sendRequest = true)
     {
         $url = $generator->svg();
-        $response = Http::get($url);
 
         $this->assertInstanceOf(DependencyUrl::class, $generator);
         $this->assertInstanceOf(DependencySvg::class, $generator);
         $this->assertStringContainsString($package, $url);
         $this->assertStringContainsString('travis-ci.com', $url);
         $this->assertStringContainsString('.svg?branch=master', $url);
-        $this->assertTrue($response->ok());
-        $this->assertStringContainsString('build', $response->body());
+
+        if ($sendRequest) {
+            $response = $this->sendRequest($url);
+
+            $this->assertStringContainsString('build', $response->body());
+        }
     }
 
     /**
      * @param string $package
      * @param DependencySvg $generator
+     * @param bool $sendRequest
      */
-    public function assertVersionSvg(string $package, DependencySvg $generator)
+    public function assertVersionSvg(string $package, DependencySvg $generator, bool $sendRequest = true)
     {
         $url = $generator->svg();
-        $response = Http::get($url);
 
         $this->assertInstanceOf(DependencyUrl::class, $generator);
         $this->assertInstanceOf(DependencySvg::class, $generator);
         $this->assertStringContainsString($package, $url);
         $this->assertStringContainsString('img.shields.io/', $url);
-        $this->assertTrue($response->ok());
 
-        $inString = (new StringHelpers($response->body()));
-        $this->assertTrue(
-            $inString->inString('<title>packagist: v') || $inString->inString('version')
-        );
+        if ($sendRequest) {
+            $response = $this->sendRequest($url);
+
+            $inString = (new StringHelpers($response->body()));
+            $this->assertTrue(
+                $inString->inString('<title>packagist: v') || $inString->inString('version'),
+                "The response body provided by {$url} doesn't contain 'packagist' or 'version'"
+            );
+        }
     }
 
     /**
      * @param string $package
      * @param DependencySvg $generator
+     * @param bool $sendRequest
      */
-    public function assertLastCommitSvg(string $package, DependencySvg $generator)
+    public function assertLastCommitSvg(string $package, DependencySvg $generator, bool $sendRequest = true)
     {
         $url = $generator->svg();
-        $response = Http::get($url);
 
         $this->assertInstanceOf(DependencyUrl::class, $generator);
         $this->assertInstanceOf(DependencySvg::class, $generator);
         $this->assertStringContainsString($package, $url);
         $this->assertStringContainsString('img.shields.io/github/last-commit', $url);
-        $this->assertTrue($response->ok());
-        $this->assertStringContainsString('last commit', $response->body());
+
+        if ($sendRequest) {
+            $response = $this->sendRequest($url);
+
+            $this->assertStringContainsString('last commit', $response->body());
+        }
     }
 
     /**
      * @param string $package
      * @param DependencyUrl $generator
+     * @param bool $sendRequest
      */
-    public function assertGithubUrl(string $package, DependencyUrl $generator)
+    public function assertGithubUrl(string $package, DependencyUrl $generator, bool $sendRequest = true)
     {
         $url = $generator->url();
-        $response = Http::get($url);
 
         $this->assertInstanceOf(DependencyUrl::class, $generator);
         $this->assertStringContainsString($package, $url);
         $this->assertStringContainsString('github.com', $url);
-        $this->assertTrue($response->ok());
+
+        if ($sendRequest) {
+            $response = $this->sendRequest($url);
+        }
     }
 
     /**
      * @param string $package
      * @param DependencyUrl $generator
+     * @param bool $sendRequest
      */
-    public function assertTravisUrl(string $package, DependencyUrl $generator)
+    public function assertTravisUrl(string $package, DependencyUrl $generator, bool $sendRequest = true)
     {
         $url = $generator->url();
-        $response = Http::get($url);
 
         $this->assertInstanceOf(DependencyUrl::class, $generator);
         $this->assertStringContainsString($package, $url);
         $this->assertStringContainsString('travis-ci.com', $url);
-        $this->assertTrue($response->ok());
+
+        if ($sendRequest) {
+            $response = $this->sendRequest($url);
+        }
     }
 
     /**
      * @param string $package
      * @param DependencyUrl $generator
+     * @param bool $sendRequest
      */
-    public function assertVersionUrl(string $package, DependencyUrl $generator)
+    public function assertVersionUrl(string $package, DependencyUrl $generator, bool $sendRequest = true)
     {
         $url = $generator->url();
-        $response = Http::get($url);
+        $inString = new StringHelpers($url);
 
         $this->assertInstanceOf(DependencyUrl::class, $generator);
-        $this->assertTrue($response->ok());
         $this->assertStringContainsString($package, $url);
-        $inString = new StringHelpers($url);
         $this->assertTrue(
-            $inString->inString('packagist.org/packages') || $inString->inString('hub.docker.com/r/')
+            $inString->inString('packagist.org/packages') || $inString->inString('hub.docker.com/r/'),
+            "The response body provided by {$url} doesn't contain 'packagist.org' or 'hub.docker.com'"
         );
+
+        if ($sendRequest) {
+            $response = $this->sendRequest($url);
+        }
+    }
+
+    /**
+     * Send an HTTP request, validate its response is "Ok" & return the response.
+     *
+     * @param string $url
+     * @return Response
+     */
+    private function sendRequest(string $url): Response
+    {
+        $response = Http::get($url);
+
+        $this->assertTrue($response->ok(), "Error: code {$response->status()} from {$url}");
+
+        return $response;
     }
 }